I32-storages are now consitently int instead of a mixture of int and long.
git-svn-id: svn://svn.code.sf.net/p/lush/code@1041 0babfeb2-de4b-461b-a2bf-1ab085750f0d

+++ lush1/trunk/src/index.c
@@ -24,7 +24,7 @@
  ***********************************************************************/
 
 /***********************************************************************
- * $Id: index.c,v 1.9 2002-04-29 20:00:35 leonb Exp $
+ * $Id: index.c,v 1.10 2002-05-02 21:08:56 leonb Exp $
  **********************************************************************/
 
 /******************************************************************************
@@ -1591,7 +1591,7 @@ void copy_index(struct index *i1, struct index *i2)
     GenericCopy(P, unsigned char);
     GenericCopy(F, flt);
     GenericCopy(D, real);
-    GenericCopy(I32, long);
+    GenericCopy(I32, int);
     GenericCopy(I16, short);
     GenericCopy(I8, char);
     GenericCopy(U8, unsigned char);
